+++ packages/react/src/components/Dropdown/Dropdown.tsx
@@ -226,6 +226,8 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
 
   static className = 'ui-dropdown'
 
+  static a11yStatusCleanupTime = 500
+
   static slotClassNames: DropdownSlotClassNames
 
   static propTypes = {
@@ -313,6 +315,12 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
       searchQuery: search ? '' : undefined,
       value: multiple ? [] : null,
     }
+  }
+
+  a11yStatusTimeout: any
+
+  componentWillUnmount() {
+    clearTimeout(this.a11yStatusTimeout)
   }
 
   public renderComponent({
@@ -954,6 +962,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
 
     if (getA11ySelectionMessage && getA11ySelectionMessage.onAdd) {
       this.setState({ a11ySelectionStatus: getA11ySelectionMessage.onAdd(item) })
+      this.setA11ySelectionMessage()
     }
 
     if (multiple) {
@@ -1038,6 +1047,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
 
     if (getA11ySelectionMessage && getA11ySelectionMessage.onRemove) {
       this.setState({ a11ySelectionStatus: getA11ySelectionMessage.onRemove(poppedItem) })
+      this.setA11ySelectionMessage()
     }
 
     this.trySetStateAndInvokeHandler('onSelectedChange', null, { value })
@@ -1132,6 +1142,17 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
 
     // otherwise, highlight no item.
     return null
+  }
+
+  /**
+   * Function that sets and cleans the selection message after it has been set,
+   * so it is not read anymore via virtual cursor.
+   */
+  private setA11ySelectionMessage = (): void => {
+    clearTimeout(this.a11yStatusTimeout)
+    this.a11yStatusTimeout = setTimeout(() => {
+      this.setState({ a11ySelectionStatus: '' })
+    }, Dropdown.a11yStatusCleanupTime)
   }
 }